requirements.txt
langchain
langchain-community
langchain-google-genai 
pillow
google-generativeai 
langchain-google-genai 
python-dotenv
faiss-cpu
cohere
streamlit
ipython